requirements.txt
Django==1.5.5
Markdown==2.3.1
South==0.8.2
argparse==1.2.1
dj-database-url==0.2.2
dj-static==0.0.5
django-filter==0.7
django-nose==1.2
django-rest-framework-docs==0.1.7
django-rest-swagger==0.1.9
django-toolbelt==0.0.1
djangorestframework==2.3.8
gunicorn==18.0
jsonpickle==0.6.1
nose==1.3.0
psycopg2==2.5.1
static==0.4
wsgiref==0.1.2
